move all click commands to 'run_script.py'
