+++ src/e2c/__init__.py
@@ -1,4 +1,4 @@
 """Main e2c module."""
-from e2c import pdf_tools, run_script
+from e2c import run_script
 
-__all__ = ["run_script", "pdf_tools"]
+__all__ = ["run_script"]
